Add files via upload
This is the codes that I make to create a mobile application for the blind for my final year project. As this is the app for the blind, there is many functions that I used involves of audio and button that the blind user just need to tap the screen of the smartphone.